Enabled a disabled test
Moved the test to a new file to add `@Tag`.
