+++ src/test/java/org/apache/ibatis/jdbc/PooledDataSourceTest.java
@@ -18,18 +18,13 @@ package org.apache.ibatis.jdbc;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
-import java.util.concurrent.TimeUnit;
 
 import org.apache.ibatis.BaseDataTest;
 import org.apache.ibatis.datasource.pooled.PooledDataSource;
 import org.hsqldb.jdbc.JDBCConnection;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 class PooledDataSourceTest extends BaseDataTest {
@@ -89,59 +84,5 @@ class PooledDataSourceTest extends BaseDataTest {
     Connection c = ds.getConnection();
     JDBCConnection realConnection = (JDBCConnection) PooledDataSource.unwrapConnection(c);
     c.close();
-  }
-
-  @Disabled("See the comments")
-  @Test
-  void shouldReconnectWhenServerKilledLeakedConnection() throws Exception {
-    // See #748
-    // Requirements:
-    // 1. MySQL JDBC driver dependency.
-    // 2. MySQL server instance with the following.
-    //  - CREATE DATABASE `test`;
-    //  - SET GLOBAL wait_timeout=3;
-    // 3. Tweak the connection info below.
-    final String URL = "jdbc:mysql://localhost:3306/test";
-    final String USERNAME = "admin";
-    final String PASSWORD = "";
-
-    PooledDataSource ds = new PooledDataSource();
-    ds.setDriver("com.mysql.jdbc.Driver");
-    ds.setUrl(URL);
-    ds.setUsername(USERNAME);
-    ds.setPassword(PASSWORD);
-    ds.setPoolMaximumActiveConnections(1);
-    ds.setPoolMaximumIdleConnections(1);
-    ds.setPoolTimeToWait(1000);
-    ds.setPoolMaximumCheckoutTime(2000);
-    ds.setPoolPingEnabled(true);
-    ds.setPoolPingQuery("select 1");
-    ds.setDefaultAutoCommit(true);
-    // MySQL wait_timeout * 1000 or less. (unit:ms)
-    ds.setPoolPingConnectionsNotUsedFor(1000);
-
-    Connection con = ds.getConnection();
-    executeQuery(con);
-    // Simulate connection leak by not closing.
-    // con.close();
-
-    // Wait for disconnected from mysql...
-    Thread.sleep(TimeUnit.SECONDS.toMillis(3));
-
-    con.close();
-
-    // Should return usable connection.
-    con = ds.getConnection();
-    executeQuery(con);
-    con.close();
-  }
-
-  private void executeQuery(Connection con) throws SQLException {
-    try (PreparedStatement st = con.prepareStatement("select 1");
-         ResultSet rs = st.executeQuery()) {
-      while (rs.next()) {
-        assertEquals(1, rs.getInt(1));
-      }
-    }
   }
 }
